refactor : dto 분리 (#7)
서비스에서 요청 dto와 응답 dto 분리
내부 클래스로 구현
