+++ src/main/java/shop/woosung/bank/dto/user/UserReqDto.java
@@ -1,0 +1,31 @@
+package shop.woosung.bank.dto.user;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import shop.woosung.bank.domain.user.User;
+import shop.woosung.bank.domain.user.UserEnum;
+
+public class UserReqDto {
+
+    @ToString
+    @Getter
+    @Setter
+    public static class SignUpReqDto {
+        private String username;
+        private String password;
+        private String email;
+        private String fullname;
+
+        public User toEntity(BCryptPasswordEncoder passwordEncoder) {
+            return User.builder()
+                    .username(username)
+                    .password(passwordEncoder.encode(password))
+                    .email(email)
+                    .fullname(fullname)
+                    .role(UserEnum.CUSTOMER)
+                    .build();
+        }
+    }
+}
